+++ catalogs/utils.ts
@@ -5,7 +5,14 @@ export function fetchIcon(url: string): Promise<string> {
     return cache.get(url)!;
   }
 
-  const promise = fetch(url).then((res) => res.text());
+  const promise = fetch(url).then((res) => {
+    if (res.ok) {
+      return res.text();
+    }
+
+    console.error(`Failed to fetch icon from ${url}`);
+    return "";
+  });
   cache.set(url, promise);
   return promise;
 }